+++ application/views/category/index.php
@@ -3,9 +3,11 @@
 
     <!-- Page Heading -->
     <h1 class="h3 mb-2 text-gray-800"><?php echo $page_title; ?></h1>
-    <a href="<?php echo base_url('category/create'); ?>" class="btn btn-primary mb-2">
-        <i class="fas fa-plus"></i> Add Category
-    </a>
+    <?php if(in_array('createCategory', $user_permission)): ?>
+        <a href="<?php echo base_url('category/create'); ?>" class="btn btn-primary mb-2">
+            <i class="fas fa-plus"></i> Add Category
+        </a>
+    <?php endif; ?>
 
     <?php if($this->session->flashdata('success')): ?>
     <div class="alert alert-success alert-dismissible" role="alert">
@@ -31,7 +33,9 @@
                         <tr>
                             <th>Name</th>
                             <th>Status</th>
-                            <th>Action</th>
+                            <?php if(in_array('updateCategory', $user_permission) || in_array('deleteCategory', $user_permission)): ?>
+                                <th>Action</th>
+                            <?php endif; ?>
                         </tr>
                     </thead>
                     <tbody>
@@ -46,20 +50,26 @@
                                 <span class="badge badge-warning">Inactive</span>
                             <?php } ?>
                             </td>
+                            <?php if(in_array('updateCategory', $user_permission) || in_array('deleteCategory', $user_permission)): ?>
                             <td>
-                                <a 
-                                    href="<?php echo base_url('category/edit/'.$data['id']); ?>"
-                                    title="edit"
-                                    class="btn btn-primary btn-sm">
-                                    <i class="fas fa-edit"></i>
-                                </a>
-                                <a 
-                                    href="#" onclick="deleteCategory('<?php echo $data['id']; ?>')"
-                                    title="delete"
-                                    class="btn btn-danger btn-sm">
-                                    <i class="fas fa-trash"></i>
-                                </a>
+                                <?php if(in_array('updateCategory', $user_permission)): ?>
+                                    <a 
+                                        href="<?php echo base_url('category/edit/'.$data['id']); ?>"
+                                        title="edit"
+                                        class="btn btn-primary btn-sm">
+                                        <i class="fas fa-edit"></i>
+                                    </a>
+                                <?php endif; ?>
+                                <?php if(in_array('deleteCategory', $user_permission)): ?>
+                                    <a 
+                                        href="#" onclick="deleteCategory('<?php echo $data['id']; ?>')"
+                                        title="delete"
+                                        class="btn btn-danger btn-sm">
+                                        <i class="fas fa-trash"></i>
+                                    </a>
+                                <?php endif; ?>
                             </td>
+                            <?php endif; ?>
                         </tr>
                         <?php } ?>
                     </tbody>
